+++ native/prime/array.hpp
@@ -12,15 +12,13 @@
 #ifndef __DTY_COMMON_NATIVE_PRIME_ARRAY_H_PLUS_PLUS__
 #define __DTY_COMMON_NATIVE_PRIME_ARRAY_H_PLUS_PLUS__
 
-#include "./iterator.hpp"
+#include "./icollections.hpp"
 
 namespace dty::collection
 {
     template<class Elem>
-    class Array final : public TianyuObject
+    class Array final : public virtual TianyuObject, public virtual ICollections<Elem>
     {
-        __PUB__ IPropertyGetter __REFERENCE__ Count = this->_Count;
-
         __PRI__ Elem  __POINTER__  _Array;
         __PRI__ int32 __POINTER__  _Reference;
 
@@ -28,7 +26,7 @@ namespace dty::collection
         __PRI__ Property<bool>  __VARIABLE__ _NeedFree;
 
         __PUB__ Array(Elem __POINTER__ arraySrc, int32 __VARIABLE__ count, bool __VARIABLE__ needFree = true)
-            : _Reference(::null), _NeedFree(needFree)
+            : TianyuObject(), ICollections<Elem>(_Count), _Reference(::null), _NeedFree(needFree)
         {
             if (::null == arraySrc && 0 != count)
                 throw dty::except::ArgumentNullException();
@@ -44,6 +42,8 @@ namespace dty::collection
             // other wise this is a counter only
             if (0 != count)
                 this->_Reference = new int32(1);
+
+            ICollections<Elem>::Count = this->_Count;
         }
         __PUB__ Array(const Array<Elem> __REFERENCE__ arr)
             : _Array(arr._Array), _Count(arr._Count), _Reference(arr._Reference), _NeedFree(arr._NeedFree)
@@ -66,7 +66,7 @@ namespace dty::collection
                 (__PTR_TO_VAR__(this->_Reference)) -= 1;
         }
 
-        __PUB__ bool           __VARIABLE__   IsNull()
+        __PUB__ virtual bool           __VARIABLE__   IsNull() override
         {
             return ::null == this->_Array;
         }
@@ -77,7 +77,7 @@ namespace dty::collection
 
             return this->_Array[index];
         }
-        __PUB__ IteratorBase<Elem> __VARIABLE__   GetIterator()
+        __PUB__ virtual Iterator<Elem> __VARIABLE__   GetIterator() override
         {
             return Iterator<Elem>(this->_Array, this->_Count);
         }
@@ -97,7 +97,7 @@ namespace dty::collection
         {
             return dty::_dty_native_cpp_default_to_string(__PTR_TO_REF__ this);
         }
-        __PUB__ virtual uint64   __VARIABLE__ GetTypeId()
+        __PUB__ virtual uint64   __VARIABLE__ GetTypeId() override
         {
             return dty::GetType(__PTR_TO_REF__ this).Id();
         }
