+++ native/prime/core/property.hpp
@@ -51,8 +51,8 @@ namespace dty
     {
         __PRI__ T __VARIABLE__ _Elem;
 
-        __PUB__ Property() : TianyuObject(), IPropertyGetter<T>(), IPropertySetter<T>(), _Elem() { }
-        __PUB__ Property(T __VARIABLE__ value) : TianyuObject(), IPropertyGetter<T>(), IPropertySetter<T>(), _Elem(value) { }
+        __PUB__ Property() : TianyuObject(), IProperty<T>(), _Elem() { }
+        __PUB__ Property(T __VARIABLE__ value) : TianyuObject(), IProperty<T>(), _Elem(value) { }
         __PUB__ virtual ~Property() override { }
 
         __PUB__ virtual ::string __VARIABLE__ ToString() noexcept override
