+++ native/prime/src/console.cpp
@@ -24,9 +24,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, ::byte __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::b2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -38,9 +38,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, ::sbyte __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::sb2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -52,9 +52,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, ::int16 __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::s2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -66,9 +66,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, ::uint16 __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::us2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -80,9 +80,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, ::int32 __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::i2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -94,9 +94,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, ::uint32 __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::ui2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -108,9 +108,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, ::int64 __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::l2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -122,9 +122,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, ::uint64 __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::ul2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -136,9 +136,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, float __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::f2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -150,9 +150,9 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, double __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
-        // newItem->_Str = ;
-        // newItem->_StrLength = ;
+
+        newItem->_Str = ::d2str(param, ::null);
+        newItem->_StrLength = ::strlen(newItem->_Str);
 
         preItem->_Next = newItem;
         newItem->_Next = ::null;
@@ -164,7 +164,6 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, const char __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
 
         newItem->_Str = new char[2]{ param, '\0' };
         newItem->_StrLength = 1;
@@ -179,7 +178,6 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, const ::string __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
 
         newItem->_StrLength = ::strlen(param);
         newItem->_Str = new char[newItem->_StrLength + 1];
@@ -197,7 +195,6 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, const ::object __VARIABLE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
 
         newItem->_StrLength = 10;
         newItem->_Str = new char[newItem->_StrLength + 1];
@@ -215,7 +212,6 @@ class ConsoleWriteHandler
         static int32 __VARIABLE__ HandleParams(OutListItem __POINTER__ preItem, T __REFERENCE__ param, Args ...params)
     {
         OutListItem __POINTER__ newItem = new OutListItem();
-        // TODO: Convert value to char array
 
         static_assert(std::is_base_of<dty::TianyuObject, T>::value, "Args...Params<T, ...> should be based from TianyuObject");
