+++ native/prime/temp/dty_collection_iterator_entity.cc
@@ -9,4 +9,4 @@
  *
  */
 
-#include "../new_iterator.hpp"
+#include "../iterator.hpp"
