+++ test/prime/all_compiler_test.cpp
@@ -2,5 +2,6 @@
 
 int main()
 {
+    dty::collection::Array<int32> a(null, 0);
     return 0;
 }
